+++ src/main/java/com/scube/chargingstation/repository/ChargingRequestRepository.java
@@ -64,10 +64,10 @@ public interface ChargingRequestRepository extends JpaRepository<ChargingRequest
 	List<ChargingRequestEntity> findByChargingPointEntity(String chargingPoint , String startDate, String endDate);
 
 	 @Query(value = "SELECT IFNULL(sum(final_kwh),0) as kwh FROM charging_request where charging_status = 'Done' and DATE_FORMAT(created_at, '%Y%m%d') = DATE_FORMAT(DATE_SUB(now(), INTERVAL 1 DAY), '%Y%m%d');", nativeQuery = true)
-	 int getYesterdayConsumedKwh();
+	 Double getYesterdayConsumedKwh();
 
 	 @Query(value = "SELECT IFNULL(sum(final_kwh),0) as kwh FROM charging_request where charging_status = 'Done' and created_at between DATE_SUB(now(), INTERVAL 7 DAY) and now();", nativeQuery = true)
-	 int getWeekConsumedKwh();
+	 Double getWeekConsumedKwh();
 
 	 @Query(value = "SELECT IFNULL(sum(cr.final_kwh),0) as kwh ,IFNULL(cp.name,'') as name FROM charging_request cr left join mst_charging_point cp on cr.fk_charging_point = cp.id group by cr.fk_charging_point order by cr.created_at , kwh;", nativeQuery = true)
 	 List<Map<String, String>> getMostActiveChargingStations();
@@ -76,7 +76,7 @@ public interface ChargingRequestRepository extends JpaRepository<ChargingRequest
 	 String get30daysTotalChargingTime();
 	 
 	 @Query(value = "SELECT IFNULL(count(id),0) AS timeSum FROM charging_request where charging_status = 'Done' and created_at between DATE_SUB(now(), INTERVAL 7 DAY) and now()", nativeQuery = true)
-	 int weekTotalChargingRequestCountSessions();
+	 Double weekTotalChargingRequestCountSessions();
 	 
 //	ChargingRequestEntity findByChargingPointEntity(ChargingPointEntity chargingPointEntity);
 
@@ -94,12 +94,12 @@ public interface ChargingRequestRepository extends JpaRepository<ChargingRequest
 	 @Query(value = "SELECT IFNULL(sum(cr.final_kwh),0) as kwh FROM charging_request cr "
 		 		+ "left join mst_charging_point cp on  cr.fk_charging_point = cp.id "
 		 		+ "where cr.charging_status = 'Done' and cp.fk_partner = (?1) and DATE_FORMAT(cr.created_at, '%Y%m%d') = DATE_FORMAT(DATE_SUB(now(), INTERVAL 1 DAY), '%Y%m%d');", nativeQuery = true)	 
-	 int getYesterdayConsumedKwhByPartnerId(String partnerid);
+	 Double getYesterdayConsumedKwhByPartnerId(String partnerid);
 	 
 	 @Query(value = "SELECT IFNULL(sum(cr.final_kwh),0) as kwh FROM charging_request cr "
 		 		+ "left join mst_charging_point cp on  cr.fk_charging_point = cp.id "
 		 		+ "where cr.charging_status = 'Done' and cp.fk_partner = (?1) and cr.created_at between DATE_SUB(now(), INTERVAL 7 DAY) and now();", nativeQuery = true)
-	 int getWeekConsumedKwhByPartnerId(String partnerid);
+	 Double getWeekConsumedKwhByPartnerId(String partnerid);
 
 	 @Query(value = "SELECT IFNULL(sum(cr.final_kwh),0) as kwh ,IFNULL(cp.name,'') as name FROM charging_request cr left join mst_charging_point cp on cr.fk_charging_point = cp.id  where  cp.fk_partner = (?1)  group by cr.fk_charging_point order by cr.created_at , kwh;", nativeQuery = true)
 	 List<Map<String, String>> getMostActiveChargingStationsByPartnerId(String partnerid);
@@ -110,7 +110,7 @@ public interface ChargingRequestRepository extends JpaRepository<ChargingRequest
 	 
 	 @Query(value = "SELECT IFNULL(count(cr.id),0) AS timeSum FROM charging_request cr left join mst_charging_point cp on  cr.fk_charging_point = cp.id "
 	 		+ " where cr.charging_status = 'Done' and cp.fk_partner = (?1) and cr.created_at between DATE_SUB(now(), INTERVAL 7 DAY) and now() ", nativeQuery = true)
-	 int weekTotalChargingRequestCountSessionsByPartnerId(String partnerid);
+	 Double weekTotalChargingRequestCountSessionsByPartnerId(String partnerid);
 	 
 }
